Add rule preprocess_snATAC
Pre-processing according to r3fang:
* Alignment filteration
* Speration of single cell
* PCR duplicates removal
* Mitochondrial reads removal
* Adjusting position of Tn5 insertion

Note: as long as we use map_scATAC this script does some redundant
things. The only reason Jee Won is using scATAC is to include the
barcode error correction. According to r3fang we would only need bwa
mem prior to snATAC pre and then do the error correction afterwards
using a script that is not implemented yet.